refactor: 코드 냄새 제거
