+++ api/src/main/java/com/garden/back/global/LocationBuilder.java
@@ -6,7 +6,7 @@ import java.net.URI;
 public class LocationBuilder {
 
     private LocationBuilder() {
-        throw new RuntimeException("유틸클래스 생성자를 통해서 객체를 생성할 수 없습니다.");
+        throw new AssertionError("유틸클래스 생성자를 통해서 객체를 생성할 수 없습니다.");
     }
 
     public static URI buildLocation(Long id) {
