+++ api/src/test/java/com/garden/back/docs/appointment/AppointmentRestDocsTest.java
@@ -23,7 +23,7 @@ import static org.springframework.restdocs.payload.PayloadDocumentation.*;
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-public class AppointmentRestDocs extends RestDocsSupport {
+class AppointmentRestDocsTest extends RestDocsSupport {
     AppointmentService appointmentService = mock(AppointmentService.class);
 
     @Override
