+++ api/src/test/java/com/garden/back/docs/feedback/FeedbackRestDocsTest.java
@@ -24,7 +24,7 @@ import static org.springframework.restdocs.request.RequestDocumentation.requestP
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-class FeedbackRestDocs extends RestDocsSupport {
+class FeedbackRestDocsTest extends RestDocsSupport {
 
     FeedbackService feedbackService = mock(FeedbackService.class);
