+++ api/src/test/java/com/garden/back/docs/post/MyPostRestDocsTest.java
@@ -29,7 +29,7 @@ import static org.springframework.restdocs.request.RequestDocumentation.queryPar
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-public class MyPostRestDocs extends RestDocsSupport {
+public class MyPostRestDocsTest extends RestDocsSupport {
     PostQueryService postQueryService = mock(PostQueryService.class);
 
     @Override
