+++ api/src/test/java/com/garden/back/docs/region/RegionRestDocsTest.java
@@ -24,7 +24,7 @@ import static org.springframework.restdocs.request.RequestDocumentation.queryPar
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-class RegionRestDocs extends RestDocsSupport {
+class RegionRestDocsTest extends RestDocsSupport {
 
     RegionService regionService = mock(RegionService.class);
