+++ chat/src/main/java/com/garden/back/util/PageMaker.java
@@ -6,7 +6,7 @@ import org.springframework.data.domain.Pageable;
 public class PageMaker {
 
     private PageMaker() {
-        throw new RuntimeException("유틸 클래스는 객체를 생성할 수 없습니다.");
+        throw new AssertionError("유틸 클래스는 객체를 생성할 수 없습니다.");
     }
 
     private static final int GARDEN_BY_NAME_PAGE_SIZE = 10;
