+++ core/src/test/java/com/garden/back/weather/infra/api/open/OpenAPIAndNaverWeatherFetcherTest.java
@@ -10,13 +10,13 @@ import com.garden.back.weather.infra.api.open.response.WeekWeatherResponse;
 import com.garden.back.weather.infra.response.AllRegionsWeatherInfo;
 import com.garden.back.weather.infra.response.WeatherPerHourAndTomorrowInfo;
 import com.garden.back.weather.infra.response.WeekWeatherInfo;
-import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
+import java.time.ZoneId;
 import java.time.format.DateTimeFormatter;
 import java.util.List;
 
@@ -106,7 +106,6 @@ class OpenAPIAndNaverWeatherFetcherTest extends IntegrationTestSupport {
 
     }
 
-    @Disabled
     @DisplayName("현재시간으로부터 1시간 단위로 5개의 날씨 정보, 다음 날 낮 12시 날씨 정보를 조회할 수 있다.")
     @Test
     void getWeatherPerHourAndTomorrowInfo() {
@@ -117,8 +116,8 @@ class OpenAPIAndNaverWeatherFetcherTest extends IntegrationTestSupport {
             .sample();
         given(naverGeoClient.getGeoInfoByLongitudeAndLatitude(anyString(), anyString(), anyString(), anyString())).willReturn(geoResponse);
 
-        String todayDate = LocalDate.now().format(DateTimeFormatter.BASIC_ISO_DATE);
-        String futureTimeFormatted = LocalTime.now().plusHours(1).format(DateTimeFormatter.ofPattern("HHmm"));
+        String todayDate = LocalDate.now(ZoneId.of("Asia/Seoul")).format(DateTimeFormatter.BASIC_ISO_DATE);
+        String futureTimeFormatted = LocalTime.now(ZoneId.of("Asia/Seoul")).plusHours(1).format(DateTimeFormatter.ofPattern("HHmm"));
 
         WeekWeatherResponse weatherResponse = sut
             .giveMeBuilder(WeekWeatherResponse.class)
